Create one instnce for each object class

+++ src/main/java/org/fasttrackit/Adopter.java
@@ -3,7 +3,7 @@ package org.fasttrackit;
 public class Adopter {
 
     String name;
-    double moneyAvailable;
+    int moneyAvailable;
